[#26] WeeklyRepeatViewModel 분리
