+++ TIG/Sources/Feature/More/EditTimeViewModel.swift
@@ -27,109 +27,28 @@ final class EditTimeViewModel {
       get { sharedState.timeSlots }
       set { sharedState.timeSlots = newValue }
     }
-    
-    var weeklyTimeSlots: [WeekDay: [TimeSlot]] = [:]
   }
   
   enum Action {
-    // EditTimeView
     case dailyTimeSaveTapped([TimeSlot])
-    
-    // WeeklyRepeatView
-    case onAppearWeeklyRepeat
-    case onChangeWeeklyTimeSlot(WeekDay, [TimeSlot])
-    case weeklyTimeSaveTapped
-    
-    // AnnounceView
-    case settingButtonTapped
   }
   
   private(set) var state: State = .init(
     sharedState: DIContainer.shared.resolve()
   )
   
-  private let appConfigRepository: AppConfigRepository = DIContainer.shared.resolve()
   private let dailyScheduleRepository: DailyScheduleRepository = DIContainer.shared.resolve()
-  private let weeklyScheduleRepository: WeeklyScheduleRepository = DIContainer.shared.resolve()
   
   func send(_ action: Action) {
     switch action {
     case .dailyTimeSaveTapped(let timeSlots):
       updateDailyTimeSlot(date: state.selectedDate, timeSlots: timeSlots)
-      
-    case .onAppearWeeklyRepeat:
-      fetchWeeklyTimeSlots()
-      
-    case .onChangeWeeklyTimeSlot(let day, let timeSlots):
-      state.weeklyTimeSlots[day] = timeSlots
-      
-    case .weeklyTimeSaveTapped:
-      updateWeeklyTimeSlots()
-    
-    case .settingButtonTapped:
-      initializeWeeklyTimeSlots()
     }
   }
 }
 
 // MARK: - Function
 private extension EditTimeViewModel {
-  /// WeeklyTimeSlots 데이터를 기본 상태로 초기화합니다.
-  func initializeWeeklyTimeSlots() {
-    
-    do {
-      let wakeup = try appConfigRepository.fetchWakeupTime().get()
-      let bed = try appConfigRepository.fetchBedTime().get()
-      let timeSlots = TimeSlot.generate(wakeup: wakeup, bed: bed)
-      
-      weeklyScheduleRepository.initializeWeeklySchedules(timeSlots: timeSlots)
-      WeekDay.allCases.forEach { state.weeklyTimeSlots[$0] = timeSlots }
-    } catch {
-      print(error)
-    }
-  }
-  
-  /// 각 요일에 해당하는 TimeSlot 데이터를 불러옵니다.
-  func fetchWeeklyTimeSlots() {
-    let result = weeklyScheduleRepository.fetchAllWeeklySchedules()
-    switch result {
-    case .success(let weeklySchedules):
-      weeklySchedules.forEach { state.weeklyTimeSlots[$0.day] = $0.timeSlots }
-    case .failure(let error):
-      print(error)
-    }
-  }
-  
-  
-  /// WeeklyTimeSlots을 업데이트 합니다.
-  func updateWeeklyTimeSlots() {
-    // 1. WeeklySchedule 업데이트
-    state.weeklyTimeSlots.forEach { weekDay, timeSlots in
-      weeklyScheduleRepository.updateWeeklySchedule(
-        weekDay: weekDay, timeSlots: timeSlots
-      )
-      
-      // 2. timeSlots이 요일에 해당된다면 업데이트
-      if state.selectedDate.weekday - 1 == weekDay.rawValue {
-        state.timeSlots = timeSlots
-      }
-      
-      do {
-        // 3. 오늘 이후에 해당되는 dailySchedule만 업데이트
-        let dailySchedules = try dailyScheduleRepository.fetchAllDailySchedules().get()
-        dailySchedules
-          .filter { $0.date >= .now.formattedDate }
-          .forEach {
-            updateDailyTimeSlot(date: $0.date, timeSlots: timeSlots)
-          }
-        
-      } catch {
-        print(error)
-      }
-    }
-  }
-  
-  
   /// DailyTimeSlots을 업데이트 합니다.
   /// - Parameters:
   ///   - date: 업데이트할 날짜.
